Fix incorrect bit-wise-and operator precedence
The compiler warns about potential precedence problem, and the compiler is
correct. The bit-wise-and operator '&' in C and C++ has a lower precedence than
the boolean=equals '==' operator, so parenenthesis is required. It turns out
that the original code accidentally produces the correct result because
`RTC_INIT_FLAG` happens to be 0x01. Any other value would have produced an
incorrect result.