fix compiler warning; fix non-atomic reading of RTC counter (#6)
* Fix incorrect bit-wise-and operator precedence

The compiler warns about potential precedence problem, and the compiler is
correct. The bit-wise-and operator '&' in C and C++ has a lower precedence than
the boolean=equals '==' operator, so parenenthesis is required. It turns out
that the original code accidentally produces the correct result because
`RTC_INIT_FLAG` happens to be 0x01. Any other value would have produced an
incorrect result.

* Handle non-atomic reading of RTC_CNTH and RTC_CNTL

The RTC_CNTH and RTC_CNTL are located in 2 separate registers so they cannot
both be read atomically. There is a small probability that the lower 16-bit
counter RTC_CNTL could rollover between the 2 reads. When that happens, the
getTime() result will be off by 65535 seconds, in either forward or backwards
direction, depending on which of the 2 registers was read first. If RTC_CNTH
and RTC_CNTL appear within a single expression, I believe the order of
evaluation is unspecified in the C/C++ language.

We fix the problem by reading the high-word (RTC_CNTH) twice, then checking if
a rollover happened between the 2 reads. If so, then the low-word (RTC_CNTL) is
read again to get the new value. The low-word will most likely be a 0x0000, but
we cannot be 100% sure of that because an interrupt could have happened right
before reading the low-word.